Speed up zp_kb.py by several orders of magnitude
This commit optimizes the zp_kb.py script, to the point where the
majority of running time is spent downloading data from zfin.

1. Previously, assigning a ZFIN EQ string to a ZP URI used an O(n^2)
algorithm. For a given URI, the script would iterate through every row
in the dataframe to find matching ZP URIs, and then take the ZFIN EQ
string from the first matching row. (They are all the same).

This commit iterates through the dataframe once and indexes ZFIN EQ
strings by ZP URI in one go.

2. This commit also changes the use of `.iterrows` to `.itertuples` when
iterating over the rows of the dataframe to avoid needlessly creating a
`Series` object.

3. Finally, this commit gets rid of two slow spots, where a string was
being concatenated for every row in the dataframe (which created lots of
needless copies of strings). Instead, output is progressively written to
the outfile.